+++ src/layouts/Disease-List/index.js
@@ -31,7 +31,7 @@ import DashboardLayout from "examples/LayoutContainers/DashboardLayout";
 import DashboardNavbar from "examples/Navbars/DashboardNavbar";
 import Footer from "examples/Footer";
 
-function Notifications() {
+function DiseaseList() {
   const [successSB, setSuccessSB] = useState(false);
   const [infoSB, setInfoSB] = useState(false);
   const [warningSB, setWarningSB] = useState(false);
@@ -194,4 +194,4 @@ function Notifications() {
   );
 }
 
-export default Notifications;
+export default DiseaseList;